requirements.txt
# Core app
streamlit==1.31.1

# OCR
easyocr==1.7.1
opencv-python-headless==4.9.0.80
pillow==10.2.0

# Audio (ASR)
whisper-openai==1.0.0
soundfile==0.12.1

# Math & utilities
numpy==1.26.4
sympy==1.12
regex==2023.12.25

# RAG & embeddings (PINNED to avoid cached_download error)
sentence-transformers==2.2.2
huggingface-hub==0.20.3

# Vector store
faiss-cpu==1.7.4

# General
scikit-learn==1.4.0
tqdm==4.66.1
